Move library to a section on the home page
Instead of a separate /library page, the library is now the bottom
section of the home page. Updated the Matt Beard article note to
"choose the games you're playing, because those games will shape you."
