+++ src/lib/library.ts
@@ -15,7 +15,7 @@ export const libraryItems: LibraryItem[] = [
 		type: 'article',
 		url: 'https://thatvastvariety.substack.com/p/your-work-will-change-you-whether',
 		publication: 'That Vast Variety',
-		note: 'Excellent commentary on choosing the games you\'re playing, and being totally aware that the games you play shape who you become.',
+		note: 'Choose the games you\'re playing, because those games will shape you.',
 		dateAdded: '2026-03-25'
 	}
 ];
