+++ src/routes/+page.js
@@ -1,5 +1,6 @@
 import { getAllProjects } from '$lib/projects';
 import { getAllPosts } from '$lib/posts';
+import { libraryItems } from '$lib/library';
 
 // since there's no dynamic data here, we can prerender
 // it so that it gets served as a static asset in production
@@ -8,5 +9,5 @@ export const prerender = true;
 export async function load() {
 	const projects = await getAllProjects();
 	const posts = await getAllPosts();
-	return { projects, posts };
+	return { projects, posts, libraryItems };
 }
